+++ dev/merkledag/viz.clj
@@ -1,7 +1,8 @@
 (ns merkledag.viz
   "Utilities for visualizing merkledag data webs."
   (:require
-    [merkledag.core :as merkle]
+    [merkledag.link :as link]
+    [merkledag.node :as node]
     [multihash.core :as multihash]
     [rhizome.viz :as rhizome]))
 
@@ -9,20 +10,20 @@
 (defn visualize-nodes
   "Constructs a graph visualizing the given collection of nodes."
   [nodes]
-  (let [node-map (->> nodes (filter :id) (map (juxt :id identity)) (into {}))]
+  (let [node-map (->> nodes (filter ::node/id) (map (juxt ::node/id identity)) (into {}))]
     (rhizome/view-graph
       (vals node-map)
       (fn adjacent [node]
-        (keep (comp node-map :target) (:links node)))
+        (keep (comp node-map ::link/target) (::node/links node)))
       :node->descriptor
         (fn [node]
-          {:label (let [short-hash (some-> (multihash/base58 (:id node)) (subs 0 8))]
-                    (if-let [data (:data node)]
-                      (remove nil? [short-hash (:data/type data) (:title data)])
+          {:label (let [short-hash (some-> (multihash/base58 (::node/id node)) (subs 0 8))]
+                    (if-let [data (::node/data node)]
+                      (remove nil? [short-hash (:data/type data) (:data/title data)])
                       short-hash))})
       :edge->descriptor
         (fn [from to]
-          (let [lname (some #(when (= (:id to) (:target %)) (:name %)) (:links from))]
+          (let [lname (some #(when (= (::node/id to) (::link/target %)) (::link/name %)) (::node/links from))]
             {:label lname})))))